pyproject.toml
[project]
name = "webapp"
version = "0.1.0"
description = "Assignment for RenewMe's Full-Stack intern position"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.124.4",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "weasyprint>=67.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
lint.ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "TCH003" # move import into type-checking block
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ARG", "S101"]
"alembic/env.py" = ["E402"]  # Allow imports after code in alembic env.py

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
disable_error_code = ["attr-defined"]
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]
exclude = ["alembic/"]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "scripts"]
skips = ["B101"]  # B101: assert statements
